+++ CHANGELOG.md
@@ -30,3 +30,7 @@
 ## Fifth Commit
 - Updated BaseRepository and BaseRepositoryInterface classes for order by and removed unnecessary code
 - Updated .env.example and added sample for all required environment variables
+
+## Sixth Commit
+- Updated routes in README.md
+- Added filter,search,order in order API